# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Gestión de Pedidos de Aberturas",
        "description": "API REST para gestionar los pedidos de una empresa de aberturas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/materials": {
            "get": {
                "tags": [
                    "Materiales"
                ],
                "summary": "Listar materiales",
                "description": "Devuelve los materiales activos ordenados por nombre. No requiere token.",
                "operationId": "0821961cdcf7173ea92453089dd0a172",
                "responses": {
                    "200": {
                        "description": "Materiales obtenidos correctamente."
                    }
                }
            },
            "post": {
                "tags": [
                    "Materiales"
                ],
                "summary": "Crear material",
                "description": "Crea un material nuevo. Solo administradores.",
                "operationId": "6cd77951b1448e367c3069ed6687379c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Acero inoxidable",
                                        "maxLength": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Material creado correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    },
                    "403": {
                        "description": "No tenés permiso para realizar esta acción."
                    },
                    "422": {
                        "description": "Los datos proporcionados no son válidos."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/materials/{id}": {
            "put": {
                "tags": [
                    "Materiales"
                ],
                "summary": "Editar material",
                "description": "Cambia el nombre de un material o lo vuelve a activar. Solo administradores.",
                "operationId": "182f62eb940ca7d446499048c5ef7047",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del material",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Aluminio anodizado",
                                        "maxLength": 100
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Material actualizado correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    },
                    "403": {
                        "description": "No tenés permiso para realizar esta acción."
                    },
                    "404": {
                        "description": "Material no encontrado."
                    },
                    "422": {
                        "description": "Los datos proporcionados no son válidos."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Materiales"
                ],
                "summary": "Desactivar material",
                "description": "Desactiva un material (no lo borra). Solo administradores.",
                "operationId": "0fa788a436a67a78b74e4da4e847c697",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del material",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material desactivado correctamente."
                    },
                    "401": {
                        "description": "No estás autenticado."
                    },
                    "403": {
                        "description": "No tenés permiso para realizar esta acción."
                    },
                    "404": {
                        "description": "Material no encontrado."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación generado por Laravel Sanctum.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}